% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLIPSE_ROSE.R
\name{add_region_signal}
\alias{add_region_signal}
\title{Get region signal from BAM files and add to GRanges}
\usage{
add_region_signal(
  sample.bam,
  regions,
  control.bam = NULL,
  floor = 1,
  read.ext = 200
)
}
\arguments{
\item{sample.bam}{Character or BamFile object representing the signal BAM file.}

\item{regions}{Character or GRanges object representing genomic regions of interest.}

\item{control.bam}{Character or BamFile object representing the control BAM file.
Default is \code{NULL}.}

\item{floor}{Numeric value for the minimum coverage threshold to count for region.
Default is 1.}

\item{read.ext}{Numeric value for extending reads downstream. Default is 200.}
}
\value{
A GRanges object for \code{regions} with additional columns for sample and control (if provided) signal.
The metadata of the object will also contain the scaling factor for library size normalization for the \code{sample.bam}
and \code{control.bam} (if provided) in the \code{sample_mmr} and \code{control_mmr} elements.
}
\description{
Calculates the libary size-normalized coverage of reads within specified regions from a
signal BAM file with optional subtraction of a control BAM file.
Allows read extension and additional processing like setting negative coverage to zero.
}
\details{
The defaults of this function are set to so as to match the functionality
of ROSE as closely as possible.

To detail the process:
\itemize{
\item The total number of reads in the signal BAM file is calculated.
\item The signal reads are extended downstream by a specified number of bases (200 bp by default).
\item The coverage for each basepair in the regions of interest are calculated.
ROSE does this manually by calling samtools for each region, which is slow.
\item Basepairs with coverage below a specified threshold (\code{floor}, 1 by default) are removed.
\item For each region, the coverage is summed and divided by the total number of reads to get the signal.
}
}
\examples{
sample.bam <- "path/to/sample.bam"
regions <- GRanges(seqnames = "chr1", ranges = IRanges(1000, 2000))
regions <- add_region_signal(sample.bam, regions, floor = 10)
}
\author{
Jared Andrews
}
