% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_enhancers.R
\name{annotate_enhancers}
\alias{annotate_enhancers}
\title{Annotate genomic ranges using TxDb and OrgDB objects}
\usage{
annotate_enhancers(
  regions,
  peaks,
  tx.db,
  org.db,
  annotate.dist = 50000,
  promoter.dist = c(2000, 200),
  annotate.ids = "SYMBOL",
  active.genes = NULL,
  identify.active.genes = FALSE,
  omit.unknown = TRUE
)
}
\arguments{
\item{regions}{A \code{GRanges} object representing genomic regions to annotate.}

\item{peaks}{A \code{GRanges} object representing peaks to identify active promoters.}

\item{tx.db}{A \code{TxDb} object containing transcript database information.
Default is \code{NULL}.}

\item{org.db}{An \code{OrgDb} object containing organism database information.
Default is \code{NULL}.}

\item{annotate.dist}{Numeric specifying the flanking distance (in bp) around the region to annotate.
Default is 50000.}

\item{promoter.dist}{Integer vector of length 2, where first and second values are upstream/downstream
distances (in bp) from the transcription start site. Used to specify the promoter region for annotations
and determining active genes.
Default is \code{c(2000, 200)}.}

\item{annotate.ids}{String specifying the gene identifier for annotations.
Default is \code{SYMBOL} (\code{ENTREZID} and \code{ENSEMBL} are not yet supported).}

\item{active.genes}{Character vector of gene \emph{symbols} to retain in the annotations.
Default is \code{NULL}.}

\item{identify.active.genes}{Logical indicating whether active genes should be identified based on
overlaps with \code{peaks}. Cannot be used simultaneously with \code{active.genes}.
Default is \code{FALSE}.}

\item{omit.unkown}{Logical indicating whether uncharacterized genes (i.e., gene symbols starting with \emph{LOC})
should be excluded from annotations.
Default is \code{TRUE}.}
}
\value{
A \code{GRanges} object with additional metadata columns for annotations.
}
\description{
Annotates a \code{GRanges} object with gene information provided by a \code{TxDb} and \code{OrgDb} object.
It can also filter for active genes or remove uncharacterized genes.
}
\details{
Annotation is performed using \code{detailRanges} from the \code{csaw} package, which
uses information provided by a \code{TxDb} and \code{OrgDb} object to identify
genes that either overlap the specified regions or are located within
a user-defined distance around each region. To filter the annotations,
users can specify a vector of active genes (symbols) or identify them
automatically based on overlaps with peaks and promoters (\code{TxDb}). Both
unfiltered and filtered annotations are appended to the metadata.
}
\examples{
# Example usage
\dontrun{
annotated.regions <- annotate_enhancers(regions = my.enhancers, peaks = my.peaks, tx.db = TxDb,
                                        org.db = OrgDb, identify.active.genes = TRUE,
                                        omit.unknown = TRUE)
}
}
\author{
Nicolas Peterson
}
