% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\name{calculate_ccf}
\alias{calculate_ccf}
\title{Function to calculate CCFs from vaf, purity and cn
Equation from NEJM 2017 paper}
\usage{
calculate_ccf(vaf, purity, tumour_totalCN, normal_totalCN, multiplicity)
}
\description{
Function to calculate CCFs from vaf, purity and cn
Equation from NEJM 2017 paper
}
