% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\name{calculate_purity}
\alias{calculate_purity}
\title{function to estimate the purity using ccf equation for clonal variants
This uses a rearrangement of the equation to calculate ccfs which is validate
for mutations we know have a ccf of 1 (those we think are clonal)}
\usage{
calculate_purity(vaf, tumour_totalCN, normal_totalCN, multiplicity)
}
\description{
function to estimate the purity using ccf equation for clonal variants
This uses a rearrangement of the equation to calculate ccfs which is validate
for mutations we know have a ccf of 1 (those we think are clonal)
}
