% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\name{call_clonal}
\alias{call_clonal}
\title{Function to assess whether subclones are in 100\% of cells or not in a given sample
ie are th ccfs significantly different from the ccfs of the clonal mtuations}
\usage{
call_clonal(
  data.,
  data_col = "ccf",
  filter_col = "hard_filtered",
  clone_col = "clone",
  clonal_col = "is_clonal",
  p_thsehold = 0.05
)
}
\description{
Function to assess whether subclones are in 100\% of cells or not in a given sample
ie are th ccfs significantly different from the ccfs of the clonal mtuations
}
