% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLIPSE_ROSE.R
\name{classify_enhancers}
\alias{classify_enhancers}
\title{Classify enhancers based on signal thresholds}
\usage{
classify_enhancers(
  regions,
  transformation = NULL,
  drop.zeros = FALSE,
  thresh.method = "ROSE",
  first.threshold = 0.5,
  arbitrary.threshold = 0.4
)
}
\arguments{
\item{regions}{A GRanges object containing \code{rank_signal} and optionally other metadata.}

\item{transformation}{A function to apply to the ranking signal before threshold determination.
Default is \code{NULL}.}

\item{drop.zeros}{Logical indicating whether to drop regions with zero signal.
Default is \code{FALSE}.}

\item{thresh.method}{Character specifying the method to determine the signal threshold.
Must be one of "ROSE", "first", "curvature", or "arbitrary".
Default is "ROSE".}

\item{first.threshold}{Numeric value for the fraction of steepest slope when using the "first" threshold method.
Higher values will result in fewer SEs called.
Default is 0.5.}

\item{arbitrary.threshold}{Numeric value for the arbitrary threshold if the "arbitrary" method is selected.
Default is 0.4, which is a reasonable setting when a cumulative proportion of signal transformation is applied.}
}
\value{
A GRanges object with a new \code{super} logical column indicating whether the enhancer is classified as a super enhancer.
A \code{rankby_signal} column is added as the final signal values used for ranking (post-transformation, if applied).
Any transformations applied, the thresholding method used, the threshold, and the number of dropped regions if
\code{drop.zeros = TRUE} are added to the metadata of the GRanges object.
}
\description{
Classifies enhancers as super enhancers or typical enhancers based on a ranking signal and
a specified thresholding method.
Optionally applies user-transformations to the signal before classification, which is highly recommended
to ameliorate the effects of outliers on the classification.
}
\examples{
regions <- GRanges(seqnames = "chr1", ranges = IRanges(1000, 2000))
regions$rank_signal <- rnorm(length(regions))
classified_regions <- classify_enhancers(regions, thresh.method = "ROSE")
}
\author{
Jared Andrews
}
