% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\name{clonal_deconvolution}
\alias{clonal_deconvolution}
\title{Clonal deconvolution function}
\usage{
clonal_deconvolution(
  data,
  normalisedSD_max = 0.56,
  sample_id_col = "sample_id",
  niose_col = "background_error",
  chromosome_col = "chr",
  position_col = "pos",
  alt_base_col = "alt",
  varcount_col = "supporting_reads",
  depth_col = "depth",
  clone_col = "clone",
  is_clonal_col = "is_clonal",
  tumour_vaf_col = "tumour_vaf",
  tumour_totcn_col = "total_cpn",
  normal_totcn_col = "normal_cpn",
  tumour_purity = "tumour_cellularity",
  tumour_ccf_col = "tumour_ccf",
  hard_filtered_col = NA,
  mrd_filtered_col = NA,
  multiplicity_col = NA,
  testing = FALSE
)
}
\description{
Clonal deconvolution function
}
