% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\name{correct_new_CIN}
\alias{correct_new_CIN}
\title{Function to detect if a large number of mutations in a clone have undergone
copy number changes so that our ccf estimates would no longer be valid
This function will try to estimate the new CN states for each mutation
Sometimes, if mutation losses have occurred but not in all cells this
function can also estimate the ccf of the new clone which has emerged
with this CIN.}
\usage{
correct_new_CIN(
  data.,
  varcount_col = "supporting_reads",
  depth_col = "depth",
  vaf_col = "vaf_nobackground",
  tumour_totalCN_col = "total_cpn",
  multimodal_p_thes = 0.05,
  background_reads_col = "background_error",
  filter_col = "hard_filtered",
  multiplicity_col = "multiplicity",
  normal_totcn_col = "normal_cpn"
)
}
\description{
Function to detect if a large number of mutations in a clone have undergone
copy number changes so that our ccf estimates would no longer be valid
This function will try to estimate the new CN states for each mutation
Sometimes, if mutation losses have occurred but not in all cells this
function can also estimate the ccf of the new clone which has emerged
with this CIN.
}
