% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialEnhancers.R
\name{differentialEnhancers}
\alias{differentialEnhancers}
\title{Perform differential analysis of super enhancers between conditions}
\usage{
differentialEnhancers(
  se.1,
  se.2,
  bam.files,
  conditions,
  window.size = 150,
  window.spacing = 50,
  paired = NULL,
  read.length = NULL,
  quality = 20,
  bg.bin.width = 2000,
  bg.fc = 3,
  merged.adj.width = 100,
  merged.max.width = 50000,
  bpparam = SerialParam(),
  debug = FALSE
)
}
\arguments{
\item{se.1}{A \code{GRanges} object of ROSE-identified enhancers for the first condition}

\item{se.2}{A \code{GRanges} object of ROSE-identified enhancers for the second condition}

\item{bam.files}{A character vector of BAM file paths that correspond to each sample.}

\item{conditions}{A character or factor vector indicating the experimental condition for \emph{each} BAM file.}

\item{window.size}{An integer specifying the window width (bp).
Default is \code{150}.}

\item{window.spacing}{An integer specifying the distance (bp) between windows.
Default is \code{50}.}

\item{paired}{A boolean specifying whether reads are paired-end.
Default is \code{NULL}.}

\item{read.length}{An integer specifying the read length. If paired-end, this argument is ignored.
If \code{NULL}, calculated using \code{correlateReads} from \code{csaw}. Can also be specified by the user.
Default is \code{NULL}.}

\item{quality}{An integer specifying the minimum mapping quality score to include reads.
Default is \code{20}.}

\item{bg.bin.width}{An integer specifying the bin width (bp) to calculate background abundance.
Default is \code{2000}.}

\item{bg.fc}{A numeric specifying the minimum fold change above background abundance required to keep windows.
Default is \code{3}.}

\item{merged.adj.width}{An integer specifying the maximum distance between adjacent windows.
Adjacent windows within this distance will be merged.
Default is \code{100}.}

\item{merged.max.width}{An integer specifying the maximum width (bp) of merged windows.
Default is \code{50000}}

\item{bpparam}{A \linkS4class{BiocParallelParam} specifying parallelization strategy.
Default is \linkS4class{SerialParam}.}

\item{debug}{Boolean specifying whether to return a list of intermediate objects for key analysis steps.
Default is \code{FALSE}.}
}
\value{
A \code{GRanges} object containing merged super enhancer regions with associated
differential analysis results including log-fold changes, p-values, FDR, and other statistics.
}
\description{
This function identifies differential signal between two conditions
using a negative binomial quasi-likelihood model. It processes enhancer regions,
performs window-based read counting, normalization, filtering based on
background signals, statistical testing, and merging of windows into
consolidated regions.
}
\details{
The analysis proceeds through the following steps:
\enumerate{
\item Provided super enhancer (\code{GRanges}) regions are concatenated to form a consensus set.
\item Consensus regions are segmented into sliding windows.
\item Reads from BAM files are counted in each window.
\item Windows with low read counts relative to background are filtered out.
\item Normalization is performed to account for trended bias.
\item Dispersion is modeled using a negative binomial generalized linear model with quasi-likelihood (QL).
\item Differential analysis is performed using a QL F-test to assign statistical significance to each window.
\item Windows are merged into regions based on proximity, and test statistics from individual windows are combined.
\item Differential analysis statistics are added to the metadata of merged regions.
}
}
\examples{
\dontrun{
significant.regions <- differentialEnhancers(se.1 = ctrl.enhancers,
                                             se.2 = treat.enhancers,
                                             bam.files = my.bam,
                                             conditions = ("ctrl", "ctrl", "drug", "drug")
                                             )
}

}
\author{
Nicolas Peterson
}
