% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSE.R
\name{enhancer_norm}
\alias{enhancer_norm}
\title{Normalizes the coverage of a candidate enhancer based on total reads and region size}
\usage{
enhancer_norm(candidate.enhancer, stitched.regions)
}
\arguments{
\item{candidate.enhancer}{RangedSummarizedExperiment object containing counts for ChIP and input.}

\item{stitched.regions}{GRanges object containing stitched regions.}
}
\value{
RangedSummarizedExperiment object with normalized read density as a new assay - NormReadDensity.
}
\description{
This replicates the normalization scheme from ROSE directly (sorta, in theory).
}
\details{
First it will extract the total number of reads in the bam file to calculate a scaling factor (total mapped reads/1M).
It will calculate region length, required for region density = (total reads in region/ region length).
Finally, it will compute normalized signal = read density / scaling factor.
}
\author{
Jacqueline Myers
}
