% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\name{extract_normalised_sd}
\alias{extract_normalised_sd}
\title{Function to extract expected normalised SDs using clonal mutations and higher
ctDNA fraction samples}
\usage{
extract_normalised_sd(
  data.,
  hard_filters,
  multiplicity_col = "mean_multiplicity",
  tumour_purity_col = "mean_tumour_cellularity",
  tumour_totcn_col = "total_cpn",
  varcount_col = "dao",
  depth_col = "ddp",
  normal_totcn_col = "normal_cpn",
  background_col = "tnc_error_rate",
  sample_id_col = "tracerx_id",
  tumour_ccf_col = "mean_tumour_ccf",
  is_clonal_col = "is_clonal",
  filter_col = "failed filters"
)
}
\description{
Function to extract expected normalised SDs using clonal mutations and higher
ctDNA fraction samples
}
