% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_differential.R
\name{find_differential}
\alias{find_differential}
\title{Perform differential analysis of super enhancers between conditions}
\usage{
find_differential(
  regions,
  g1.bam.files,
  g2.bam.files,
  g1.name = "group1",
  g2.name = "group2",
  window.size = 150,
  window.spacing = 50,
  param = readParam(minq = 20, dedup = FALSE),
  paired = FALSE,
  fragment.length = 200,
  bg.bin.width = 2000,
  bg.fc = 3,
  merged.adj.width = 100,
  merged.max.width = 50000,
  BPPARAM = SerialParam(),
  debug = FALSE
)
}
\arguments{
\item{regions}{A \code{GRanges} object of regions to compare between conditions.
These regions will be reduced, i.e. overlapping regions will be merged to one.}

\item{g1.bam.files}{A character vector of BAM file paths for group 1.}

\item{g2.bam.files}{A character vector of BAM file paths for group 2.}

\item{g1.name}{A character string specifying the name of group 1.
Default is "group1".}

\item{g2.name}{A character string specifying the name of group 2.
Default is "group2".}

\item{window.size}{An integer specifying the window width (bp).
Default is \code{150}.}

\item{window.spacing}{An integer specifying the distance (bp) between windows.
Default is \code{50}.}

\item{param}{A \link[csaw:readParam]{csaw::readParam} object specifying parameters for read counting.
This can be used to restrict chromosomes or exclude intervals (like blacklists).
Default is \code{readParam(minq = 20, dedup = FALSE)}.}

\item{paired}{A boolean specifying whether reads are paired-end.
Default is \code{FALSE}.}

\item{fragment.length}{An integer specifying the fragment length.
If paired-end, this argument is ignored.
If \code{NULL}, calculated using \code{\link[csaw:correlateReads]{csaw::correlateReads()}}, which is rather slow.
Default is \code{200}.}

\item{bg.bin.width}{An integer specifying the bin width (bp) to calculate background abundance.
Default is \code{2000}.}

\item{bg.fc}{A numeric specifying the minimum fold change above background
abundance required to keep windows.
Default is \code{3}.}

\item{merged.adj.width}{An integer specifying the maximum distance between adjacent windows.
Adjacent windows within this distance will be merged.
Default is \code{100}.}

\item{merged.max.width}{An integer specifying the maximum width (bp) of merged windows to mitigate
excessive daisy chaining.
Default is \code{30000}.}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} specifying parallelization strategy.
Default is \link[BiocParallel:SerialParam-class]{BiocParallel::SerialParam}.}

\item{debug}{Boolean specifying whether to return a list of intermediate objects for key analysis steps.
Default is \code{FALSE}.}
}
\value{
A named list containing:
\itemize{
\item \strong{merged_regions} - A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object containing regions with associated
differential analysis results including log-fold changes, p-values, FDR, and other statistics.
\item \strong{filtered_window_counts} - A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment}
object containing filtered window counts, useful for coverage calculation, viz, etc.
}
}
\description{
This function identifies differential signal between two conditions
using a negative binomial quasi-likelihood model. It processes enhancer regions,
performs window-based read counting, normalization, filtering based on
background signals, statistical testing, and merging of windows into
consolidated regions.
}
\details{
The analysis proceeds through the following steps:
\enumerate{
\item Regions are segmented into sliding windows.
\item Reads from BAM files are counted in each window.
\item Windows with low read counts relative to background are filtered out.
\item Normalization is performed to account for trended bias.
\item Dispersion is modeled using a negative binomial generalized linear
model with quasi-likelihood (QL).
\item Differential analysis is performed using a QL F-test to
assign statistical significance to each window.
Group 2 will be set as the reference level.
\item Windows are merged into regions based on proximity, and test statistics
from individual windows are combined via \code{\link[csaw:mergeResults]{csaw::mergeResults()}}.
\item Differential analysis statistics are added to the metadata of merged regions.
}
}
\examples{
\dontrun{
significant.regions <- find_differential(
    regions = my.granges,
    g1.bam.files = my.bam1,
    g2.bam.files = my.bam2
)
}

}
\author{
Nicolas Peterson, Jared Andrews
}
