% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLIPSE_ROSE.R
\name{get_ranking_signal}
\alias{get_ranking_signal}
\title{Get ranking signal for regions}
\usage{
get_ranking_signal(regions, negative.to.zero = TRUE)
}
\arguments{
\item{regions}{A GRanges object containing the \code{sample_signal} and optionally \code{control_signal} in metadata columns.}

\item{negative.to.zero}{Logical indicating whether to set negative values in the ranking signal to zero.
Default is \code{TRUE}.}
}
\value{
A GRanges object with an added \code{rank_signal} column containing the
computed \code{rank_signal} column in its metadata columns, sorted by said column.
Adds a \code{region_rank} column as well.
}
\description{
Computes the ranking signal for the given genomic regions by
optionally subtracting control signal and setting negative values to zero.
}
\examples{
regions <- GRanges(seqnames = "chr1", ranges = IRanges(1000, 2000))
regions$sample_signal <- rnorm(length(regions))
regions$control_signal <- rnorm(length(regions))
ranked_regions <- get_ranking_signal(regions)
}
\author{
Jared Andrews, Jacqueline Myers
}
