% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\name{outlier_test}
\alias{outlier_test}
\title{Function to identify outliers until SD is acceptable and identify clones which
have an incoherent vaf distribution (ie probably not a true clone) because the normalised SD is still
too high after removing 4 outliers (or if >50\% of mutations are removed as outliers)
The normalisedSD_max is claculated by applying the extract_normalised_sd to the whole cohort}
\usage{
outlier_test(
  data.,
  normalisedSD_max,
  data_col = "cn_adj_vaf",
  LOD_col = "cn_adj_vaf_LOD",
  filter_col = "hard_filtered",
  outlier_perc_limit = 0.5,
  outlier_num_limit = 4
)
}
\description{
Function to identify outliers until SD is acceptable and identify clones which
have an incoherent vaf distribution (ie probably not a true clone) because the normalised SD is still
too high after removing 4 outliers (or if >50\% of mutations are removed as outliers)
The normalisedSD_max is claculated by applying the extract_normalised_sd to the whole cohort
}
