% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_enhancer_curve}
\alias{plot_enhancer_curve}
\title{Plot enhancer ranking curve}
\usage{
plot_enhancer_curve(regions, style = "ROSE", factor_label = NULL)
}
\arguments{
\item{regions}{A GRanges object containing \code{rankby_signal} and \code{super} columns, as well as the \code{threshold} value in its metadata.}

\item{style}{Character string specifying the plotting style.
Default is "ROSE".}

\item{factor_label}{Optional character string for prefixing the y-axis label with a specific factor,
e.g. "(H3K27ac)".
Default is \code{NULL}.}
}
\value{
None. This function generates a plot.
}
\description{
Plots the enhancer ranking curve based on the signal values in the regions,
with visual indicators for the signal cutoff and the number of super-enhancers identified.
}
\examples{
regions <- GRanges(seqnames = "chr1", ranges = IRanges(1000, 2000))
regions$rankby_signal <- rnorm(length(regions))
regions$super <- regions$rankby_signal > 1
metadata(regions)$threshold <- 1
plot_enhancer_curve(regions)
}
\author{
Jared Andrews, based on code from ROSE.
}
