% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_locus}
\alias{plot_locus}
\title{Plot Locus}
\usage{
plot_locus(
  region.gr,
  window.gr,
  window.rse,
  g1.bam.files,
  g2.bam.files,
  genome = "hg38",
  region.track.name = "SEs",
  region.fill = "orange",
  padding = 0.3,
  fdr.thresh = 0.05,
  fc.track.position = c("between", "above", "below"),
  g1.bw.files = NULL,
  g2.bw.files = NULL,
  g1.fill = "red",
  g2.fill = "blue",
  param = readParam(minq = 20, dedup = FALSE),
  top.tracks = NULL,
  bottom.tracks = NULL,
  main = "",
  region.lfc.color = "black",
  highlight.sig = TRUE,
  higlight.fill = "#d3ff8c",
  highlight.color = "#8ff7df",
  ...
)
}
\arguments{
\item{region.gr}{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object containing a single region to plot.}

\item{window.gr}{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object representing merged windows with
associated statistics, as returned by \code{\link[=find_differential]{find_differential()}}.}

\item{window.rse}{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment} object
containing read count data and library sizes, as returned by \code{\link[=find_differential]{find_differential()}}.}

\item{g1.bam.files}{Named list of BAM file paths for group 1.}

\item{g2.bam.files}{Named list of BAM file paths for group 2.}

\item{genome}{String specifying the genome assembly.
This is generally expected to match the UCSC conventions for genome assemblies.
Default is "hg38".}

\item{region.track.name}{String for the name of the region track.
Default is "SEs".}

\item{padding}{Numeric value specifying the padding around regions.
This is used to expand the region for additional context by expanding
the region by a factor of (1 + padding).
Default is 0.3, e.g. 30\% of the region size.}

\item{fdr.thresh}{Numeric value specifying the FDR threshold for significance.
Default is 0.05.}

\item{fc.track.position}{String specifying the position of the fold-change track
relative to signal tracks. Options are "between", "above", or "below".
Default is "between".}

\item{g1.bw.files}{Optional list of BigWig file paths for group 1.
Not implemented currently.}

\item{g2.bw.files}{Optional list of BigWig file paths for group 2.
Not implemented currently.}

\item{g1.fill}{String specifying the fill color for group 1 tracks.
Default is "red".}

\item{g2.fill}{String specifying the fill color for group 2 tracks.
Default is "blue".}

\item{param}{\link[csaw:readParam]{csaw::readParam} object specifying read extraction parameters.
Default is \code{readParam(minq = 20, dedup = FALSE)}.}

\item{top.tracks}{Optional list of additional tracks to display above the main tracks.
Default is NULL.}

\item{bottom.tracks}{Optional list of additional tracks to display below the main tracks.
Default is NULL.}

\item{main}{String specifying the main title of the plot.
Default is "".}

\item{region.lfc.color}{String specifying the color of the log2 fold-change track.
Default is "black".}

\item{highlight.sig}{Logical indicating whether to highlight significant regions.
Default is TRUE.}

\item{higlight.fill}{String specifying the fill color for highlighted regions.
Default is "#d3ff8c".}

\item{highlight.color}{String specifying the border color for highlighted regions.
Default is "#8ff7df".}

\item{...}{Additional arguments passed to the \code{plotTracks} function.}
}
\value{
A list of Gviz tracks.
}
\description{
This function generates a locus visualization plot using genomic regions,
signal tracks, and fold-change data.
It supports highlighting significant
regions and customizing the appearance of tracks.
}
\details{
Coverage tracks are generated from BAM files and normalized by library size to reads per million.
While not a perfect normalization, as it doesn't account for composition effects, it is often
sufficient for visualization of robust differences between groups with sufficient numbers of replicates.
}
\examples{
# Example usage:
plot_locus(
  region.gr = my_regions,
  window.gr = my_windows,
  window.rse = my_rse,
  g1.bam.files = list("sample1" = "path/to/sample1.bam"),
  g2.bam.files = list("sample2" = "path/to/sample2.bam"),
  genome = "hg38",
  region.track.name = "My Regions",
  padding = 0.2,
  fdr.thresh = 0.01
)

}
\author{
Jared Andrews
}
