% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_qq}
\alias{plot_qq}
\title{Generate QQ Plot for a Specified Distribution}
\usage{
plot_qq(
  data,
  column,
  dist,
  point.size = 1,
  point.color = "blue",
  line.width = 0.8,
  line.color = "black",
  do.raster = FALSE,
  raster.dpi = 300,
  return.plotly = FALSE
)
}
\arguments{
\item{data}{A GRanges object containing the data to fit in \code{mcols}.}

\item{column}{A string specifying the column name in the \code{mcols(data)} to use for values.}

\item{dist}{A string specifying the distribution to fit. Supported distributions include:
"beta", "cauchy", "chi-squared", "exponential", "gamma", "geometric", "log-normal", "lognormal",
"logistic", "negative binomial", "normal", "Poisson", "t", and "weibull".}

\item{point.size}{Numeric value specifying the size of points in the QQ plot.
Default is 1.}

\item{point.color}{A string specifying the color of points in the QQ plot.
Default is "blue".}

\item{line.width}{Numeric value specifying the width of the reference line.
Default is 0.8.}

\item{line.color}{A string specifying the color of the reference line.
Default is "red".}

\item{do.raster}{Logical indicating whether to render the plot as a raster image.
Default is \code{FALSE}.}

\item{raster.dpi}{Numeric value specifying the DPI of the raster image if \code{do.raster} is \code{TRUE}.}

\item{return.plotly}{Logical indicating whether to return a plotly object.
Default is \code{FALSE}.}
}
\value{
A ggplot2 object.
}
\description{
Generate QQ Plot for a Specified Distribution
}
\examples{
gr <- GRanges(
    seqnames = rep("chr1", 1000),
    ranges = IRanges(start = seq(1, by = 1000, length.out = 1000), end = seq(1000, by = 1000, length.out = 1000))
)
mcols(gr) <- DataFrame(values = rnorm(1000))
plot_qq(data = gr, column = "values", dist = "normal")
plot_qq(data = gr, column = "values", dist = "gamma")
}
\author{
Jared Andrews
}
