% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\name{power_calc}
\alias{power_calc}
\title{Function to estimate our power to significantly detect either sample purity (type = 'power_purity')
clone ccf (type = 'power_clone_ccf') or the ccf of the average subclone in a sample (type = 'power_sample_ccf')
based on the background noise, copy number status, depth of sequencing etc. Default is for detection at
p = 0.01}
\usage{
power_calc(
  data.,
  type,
  niose_col = "background_error",
  filter_col = "mrd_filtered",
  normal_cn_col = "normal_cn",
  purity_col = "purity",
  multiplicity_col = "multiplicity",
  tumour_totcn_col = "total_cpn",
  depth_col = "depth",
  clonal_col = "is_clonal",
  p_theshold = 0.05,
  num_subclonal_mutations_sim = 5
)
}
\description{
Function to estimate our power to significantly detect either sample purity (type = 'power_purity')
clone ccf (type = 'power_clone_ccf') or the ccf of the average subclone in a sample (type = 'power_sample_ccf')
based on the background noise, copy number status, depth of sequencing etc. Default is for detection at
p = 0.01
}
