% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\name{remove_background_reads}
\alias{remove_background_reads}
\title{Function to remove the number of estimated supporting reads which derive from
background sequencing niose for groups of mutations (in this case clones)}
\usage{
remove_background_reads(
  data.,
  niose_col = "background_error",
  filter_col = "hard_filtered",
  varcount_col = "supporting_reads",
  depth_col = "depth",
  group_max = 4,
  mutid_col = "mutation_id"
)
}
\description{
Function to remove the number of estimated supporting reads which derive from
background sequencing niose for groups of mutations (in this case clones)
}
