% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLIPSE_ROSE.R
\name{run_rose}
\alias{run_rose}
\title{Run ROSE (Rank Ordering of Super-Enhancers)}
\usage{
run_rose(
  sample.bam,
  peaks,
  control.bam = NULL,
  stitch.distance = 12500,
  txdb = NULL,
  org.db = NULL,
  drop.y = TRUE,
  max.genes.overlap = 2,
  negative.to.zero = TRUE,
  thresh.method = "ROSE",
  transformation = NULL,
  floor = 1,
  read.ext = 200,
  drop.zeros = FALSE,
  first.threshold = 0.5,
  arbitrary.threshold = 0.4
)
}
\arguments{
\item{sample.bam}{A character string or `BamFile`` object representing the sample BAM file.}

\item{peaks}{A character string or `GRanges`` object representing the peaks.}

\item{control.bam}{A character string or \verb{BamFile`` object representing the control BAM file. Default is }NULL`.}

\item{stitch.distance}{Numeric value for the distance within which peaks are stitched together.
Default is \code{12500}.}

\item{txdb}{Not used. Functionality not implemented.
Default is \code{NULL}.}

\item{org.db}{Not used. Functionality not implemented.
Default is \code{NULL}.}

\item{drop.y}{Logical indicating whether to drop peaks on chromosome Y.
Default is \code{TRUE}.}

\item{max.genes.overlap}{Not used. Functionality not implemented.
Default is 2.}

\item{negative.to.zero}{Logical indicating whether to set negative ranking signals to zero.
Default is \code{TRUE}.}

\item{thresh.method}{Character string specifying the method to determine the signal threshold.
Must be one of "ROSE", "first", "curvature", or "arbitrary".
Default is "ROSE".}

\item{transformation}{A function to apply to the ranking signal before threshold determination.
Default is \code{NULL}.}

\item{floor}{Numeric value representing the minimum coverage threshold to count.
Default is \code{1}.}

\item{read.ext}{Numeric value for extending reads downstream.
Default is \code{200}.}

\item{drop.zeros}{Logical indicating whether to drop regions with zero signal.
Default is \code{FALSE}.}

\item{first.threshold}{Numeric value for the fraction of steepest slope when using the "first" threshold method.
Default is \code{0.5}.}

\item{arbitrary.threshold}{Numeric value for the arbitrary threshold if the "arbitrary" method is selected.
Default is \code{0.4}.}
}
\value{
A GRanges object containing the classified super-enhancers and associated metadata.
}
\description{
This function performs the ROSE for identifying super-enhancers by stitching
together peaks, calculating the signal in the regions, ranking the regions by signal,
and classifying them as super-enhancers.
}
\details{
This function allows for near identical functionality to the original ROSE software,
but also provides a number of additional options to minimize the impact of signal outliers
on the classification threshold.
In particular, the \code{transformation} argument allows for
the application of a function to the ranking signal before threshold determination.
This is highly recommended to prevent outliers from skewing the threshold.

The \code{thresh.method} argument allows for the selection of the method to determine the threshold.
\itemize{
\item The "ROSE" method is the default and uses a sliding diagonal line to determine the cutoff.
\item The "first" method uses a first derivative to determine the point where the slope is a given fraction of the maximum.
The \code{first.threshold} argument controls this fraction.
\item The "curvature" method finds the point at which the circle tanget to the curve has the smallest radius.
\item The "arbitrary" method allows for the user to specify a fixed threshold, useful for transformations
that result in a consistent curve shape with a known maximum, like cumulative proportion of signal.
}
}
\examples{
sample.bam <- "path/to/sample.bam"
peaks <- "path/to/peaks.bed"
result <- run_rose(sample.bam, peaks)
}
\author{
Jared Andrews
}
