% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLIPSE_ROSE.R
\name{unstitch_regions}
\alias{unstitch_regions}
\title{Unstitch regions based on overlaps with TSS}
\usage{
unstitch_regions(stitched, original, tss, id.col = "GENEID", threshold = 2)
}
\arguments{
\item{stitched}{A `GRanges`` object representing the stitched regions.}

\item{original}{A `GRanges`` object representing the original regions prior to stitching.}

\item{tss}{A `GRanges`` object representing the TSS regions.}

\item{id.col}{A character string specifying the column name in the TSS object that contains gene IDs.
Default is "GENEID".}

\item{threshold}{An integer specifying the threshold for the number of unique genes before unstitching is applied.
Default is 2.}
}
\value{
A list with a \code{hits} data.frame with the following columns:
\describe{
\item{qhits}{Query hits corresponding to the stitched regions.}
\item{chrom}{Chromosome names of the stitched regions.}
\item{start}{Start positions of the stitched regions.}
\item{end}{End positions of the stitched regions.}
\item{num_olap}{Number of unique genes overlapping each stitched region.}
\item{ids}{Comma-separated gene IDs overlapping each stitched region.}
}

The \code{regions} element of the list contains a GRanges object with the stitched regions
and the original regions resulting from unstitching.
}
\description{
This function takes stitched regions and identifies overlaps with TSS regions.
If a given stitched region overlaps TSS from more than a specified number of unique genes,
the region is unstitched into its original components from the \code{original} regions.
}
\details{
This attempts to emulate the method used by ROSE, but the results it returns are not identical due to a presumed bug in ROSE
wherein it does not always properly find all overlaps.
}
\examples{
library(GenomicRanges)

# Example stitched regions, original peaks, and TSSes
stitched <- GRanges(seqnames = c("chr1", "chr1", "chr2"),
                    ranges = IRanges(start = c(100, 300, 500),
                    end = c(200, 400, 600)),
                    strand = c("+", "-", "+"))

original <- GRanges(seqnames = c("chr1", "chr1", "chr2", "chr2"),
                    ranges = IRanges(start = c(100, 150, 300, 500),
                    end = c(150, 200, 350, 550)),
                    strand = c("+", "+", "-", "+"))

tss <- GRanges(seqnames = c("chr1", "chr1", "chr1", "chr1", "chr2"),
               ranges = IRanges(start = c(120, 140, 160, 320, 520), 
               end = c(130, 150, 170, 330, 530)),
               strand = c("+", "+", "+", "-", "+"),
               GENEID = c("gene1", "gene11", "gene111", "gene2", "gene3"))

result <- unstitch_regions(stitched, original, tss, id.col = "GENEID", threshold = 2)

result$regions
}
\author{
Jared Andrews
}
