% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLIPSE_ROSE.R
\name{classify_enhancers}
\alias{classify_enhancers}
\title{Classify enhancers based on signal thresholds}
\usage{
classify_enhancers(
  regions,
  transformation = NULL,
  thresh.method = "ROSE",
  rose.slope = NULL,
  first.threshold = 0.5,
  second_diff.threshold = 0,
  chord.threshold = 0,
  arbitrary.threshold = 0.4,
  segmented.breakpoints = 3,
  mads = 3
)
}
\arguments{
\item{regions}{A \code{GRanges} object containing \code{rank_signal} and optionally other metadata.}

\item{transformation}{A function to apply to the ranking signal before threshold determination.
Default is \code{NULL}.}

\item{thresh.method}{Character string specifying the method to determine the signal threshold.
Must be one of "ROSE", "first", "second_diff", "segmented", "chord", "curvature", "mad", or "arbitrary".
Default is "ROSE".}

\item{rose.slope}{Numeric value for the slope of the diagonal line used for the "ROSE" threshold method.
Default is \code{NULL}, meaning it will be calculated automatically based on the minimum and maximum signal values.
See \code{calculate_cutoff} for details.}

\item{first.threshold}{Numeric value for the fraction of steepest slope when using the "first" threshold method.
Higher value will result in fewer SEs called.
Default is 0.5.}

\item{second_diff.threshold}{Numeric value for ignoring small differences when using the "second_diff" threshold method.
May be useful for transformation that result in curves with small differences.
Generally, unlikely to be useful for most applications.
Default is 0.}

\item{chord.threshold}{Numeric value for ignoring small differences when using the "chord" threshold method.
May be useful for transformation that result in curves with small differences.
Generally, unlikely to be useful for most applications.
Default is 0.}

\item{arbitrary.threshold}{Numeric value for the arbitrary threshold if the "arbitrary" threshold method is selected.
Useful for scaled transformations like cumulative proportion of signal.
Default is 0.4.}

\item{segmented.breakpoints}{Number of breakpoints to find when using the "segmented" threshold method.
Generally, this should be set to the rough number of linear segments in the curve.
Default is 3.}

\item{mads}{Numeric value for the number of median absolute deviations to use for the "mad" threshold method.
Increasing will reduce the number of SEs called.
Default is 3.}
}
\value{
A \code{GRanges} object with a new \code{super} logical column indicating whether the enhancer is classified as a super enhancer.
A \code{rankby_signal} column is added as the final signal values used for ranking (post-transformation, if applied).
Any transformations applied, the thresholding method used, the threshold, and the number of dropped regions if
\code{drop.no.signal = TRUE} are added to the metadata of the \code{GRanges} object.
}
\description{
Classifies enhancers as super enhancers or typical enhancers based on a ranking signal and
a specified thresholding method.
Optionally applies user-transformations to the signal before classification, which is highly recommended
to ameliorate the effects of outliers on the classification.
}
\examples{
library(GenomicRanges)
regions <- GRanges(
    seqnames = c("chr1", "chr2", "chr3"),
    ranges = IRanges(
        start = c(100, 200, 300),
        end = c(200, 300, 400)
    )
)
regions$rank_signal <- rnorm(length(regions))
classified_regions <- classify_enhancers(regions, thresh.method = "ROSE")
}
\author{
Jared Andrews
}
