% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLIPSE_ROSE.R
\name{extend_reads}
\alias{extend_reads}
\title{Extend reads directionally upstream and downstream}
\usage{
extend_reads(regions, upstream = 0, downstream = 0)
}
\arguments{
\item{regions}{A \code{GRanges} object containing genomic regions.}

\item{upstream}{Number of bases to extend upstream.
Default is 0.}

\item{downstream}{Number of bases to extend downstream.
Default is 0.}
}
\value{
A \code{GRanges} object with extended regions.
}
\description{
Extends reads by a specified number of bases upstream or downstream
based on strand orientation.
}
\examples{
library(GenomicRanges)
regions <- GRanges(seqnames = "chr1", ranges = IRanges(100, 200), strand = "+")
extended <- extend_reads(regions, upstream = 50, downstream = 100)
extended
}
\author{
Jared Andrews
}
