% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSE.R
\name{fetch_tss}
\alias{fetch_tss}
\title{Fetches TSSes or promoters from a TxDb object}
\usage{
fetch_tss(TxDb, expansion.distance = 0)
}
\arguments{
\item{TxDb}{TxDb object to fetch TSSes from.}

\item{expansion.distance}{Integer \strong{distance} to add up/downstream of TSSes.}
}
\value{
GRanges object of TSSes for a TxDb object.
}
\description{
Fetches TSSes or promoters from a TxDb object
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
TxDb <- TxDb.Hsapiens.UCSC.hg19.knownGene
tsses <- fetch_tss(TxDb, expansion.distance = 1000)
}
\author{
Jacqueline Myers
}
