% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_enhancer_curve}
\alias{plot_enhancer_curve}
\title{Plot enhancer ranking curve}
\usage{
plot_enhancer_curve(regions, style = "ROSE", factor_label = NULL)
}
\arguments{
\item{regions}{A GRanges object containing \code{rankby_signal} and \code{super} columns, as well as the \code{threshold} value in its metadata.}

\item{style}{Character string specifying the plotting style.
Default is "ROSE".}

\item{factor_label}{Optional character string for prefixing the y-axis label with a specific factor,
e.g. "(H3K27ac)".
Default is \code{NULL}.}
}
\value{
None. This function generates a plot.
}
\description{
Plots the enhancer ranking curve based on the signal values in the regions,
with visual indicators for the signal cutoff and the number of super-enhancers identified.
}
\examples{
library(GenomicRanges)

set.seed(7)
num_elements <- 1000
regions <- GRanges(seqnames = rep("chr1", num_elements),
                   ranges = IRanges(start = seq(1, by = 1000, length.out = num_elements),
                                    end = seq(1000, by = 1000, length.out = num_elements)))

# Add random signal values and classify as super enhancers
regions$rankby_signal <- c(rnorm(num_elements - 50, mean = 150, sd = 10), 
                           rnorm(50, mean = 250, sd = 50))
regions$super <- regions$rankby_signal > 170
metadata(regions)$threshold <- 170

# Plot the enhancer curve
plot_enhancer_curve(regions)
}
\author{
Jared Andrews, based on code from ROSE.
}
