% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLIPSE_ROSE.R
\name{run_rose}
\alias{run_rose}
\title{Run ROSE (Rank Ordering of Super-Enhancers)}
\usage{
run_rose(
  treatment,
  peaks,
  control = NULL,
  stitch.distance = 12500,
  tss.exclusion.distance = 0,
  txdb = NULL,
  org.db = NULL,
  drop.y = TRUE,
  max.unique.gene.tss.overlap = NULL,
  tss.overlap.distance = 50,
  negative.to.zero = TRUE,
  drop.no.signal = FALSE,
  thresh.method = "ROSE",
  first.threshold = 0.5,
  arbitrary.threshold = 0.4,
  segmented.breakpoints = 3,
  transformation = NULL,
  floor = 1,
  read.ext = 200,
  normalize.by.width = FALSE,
  drop.zeros = FALSE,
  annotate = TRUE,
  annotate.dist = 50000,
  promoter.dist = c(2000, 200),
  active.genes = NULL,
  identify.active.genes = FALSE,
  omit.unknown = TRUE
)
}
\arguments{
\item{treatment}{A \code{BamFile} or \code{GRanges} object representing the sample signal.}

\item{peaks}{A  \code{GRanges} object representing the peaks.}

\item{control}{A \code{BamFile} or \code{GRanges} object representing the control (usually input or IgG) signal.
Default is \code{NULL}.}

\item{stitch.distance}{Numeric value for the distance within which peaks are stitched together.
Default is 12500.}

\item{tss.exclusion.distance}{Numeric value for distance to add to TSS exclusion range prior to stitching.
Peaks \emph{fully contained} within this exclusion range will be excluded from the stitching process.
Default is 0, meaning peaks overlapping TSS will not be excluded from stitching.}

\item{txdb}{Transcript annotations used for peak exclusion for TSS overlap and unstitching process for
regions spanning TSSes from more than \code{max.unique.gene.tss.overlap} genes. Also used for enhancer annotation.}

\item{org.db}{An \code{OrgDb} object containing organism database information. Used for enhancer annotation.
Default is \code{NULL}.}

\item{drop.y}{Logical indicating whether to drop peaks on chromosome Y, as done by the original ROSE implementation.
Default is \code{TRUE}.}

\item{max.unique.gene.tss.overlap}{Maximum number of unique genes that a region can overlap the TSS of before being unstitched.
Note that multiple overlapping regions from the same gene, e.g. multiple isoforms, are counted as one.
Default is \code{NULL}. Ignored if \code{txdb} is \code{NULL}.}

\item{tss.overlap.distance}{Numeric value for the distance (in bp) to add to TSS for unstitching.
Default is 50. Ignored if \code{txdb} is \code{NULL}.}

\item{negative.to.zero}{Logical indicating whether to set negative ranking signals to zero.
Default is \code{TRUE}.}

\item{drop.no.signal}{Logical indicating whether to remove regions with negative or zero values in the ranking signal.
Default is \code{FALSE}.}

\item{thresh.method}{Character string specifying the method to determine the signal threshold.
Must be one of "ROSE", "first", "second_diff", "segmented", "chord", "curvature", or "arbitrary".
Default is "ROSE".}

\item{first.threshold}{Numeric value for the fraction of steepest slope when using the "first" threshold method.
Default is 0.5.}

\item{arbitrary.threshold}{Numeric value for the arbitrary threshold if the "arbitrary" threshold method is selected.
Useful for scaled transformations like cumulative proportion of signal.
Default is 0.4.}

\item{segmented.breakpoints}{Number of breakpoints to find when using the "segmented" threshold method.
Generally, this should be set to the rough number of linear segments in the curve.
Default is 3.}

\item{transformation}{A function to apply to the ranking signal before threshold determination.
Default is \code{NULL}.}

\item{floor}{Numeric value representing the minimum coverage threshold to count.
Default is 1.}

\item{read.ext}{Numeric value for extending reads downstream.
Default is 200. Ignored if inputs are \code{GRanges} objects.}

\item{normalize.by.width}{Logical indicating whether to normalize signal by region width.
ROSE does this, but then undoes it at a later step, so \code{FALSE} is the default.}

\item{drop.zeros}{Logical indicating whether to drop regions with zero signal.
Default is \code{FALSE}.}

\item{annotate}{Logical indicating whether annotations should be provided.
Default is \code{TRUE}.}

\item{annotate.dist}{Numeric specifying the flanking distance (in bp) around the region to use
for annotation, i.e. genes within this distance will be associated with the region.
Default is 50000.}

\item{promoter.dist}{Integer vector of length 2, where first and second values are upstream/downstream
distances (in bp) from the transcription start site. Used to specify the promoter region for annotations
and determining active genes.
Default is \code{c(2000, 200)}.}

\item{active.genes}{Character vector of gene \emph{symbols} to retain in the annotations.
Can be useful for retaining only expressed genes.
Default is \code{NULL}.}

\item{identify.active.genes}{Logical indicating whether active genes should be identified based on
overlaps with \code{peaks}. Will be ignored if \code{active.genes} provided.
Default is \code{FALSE}.}

\item{omit.unknown}{Logical indicating whether uncharacterized genes (i.e., gene symbols starting with \emph{LOC})
should be excluded from annotations.
Default is \code{TRUE}.}

\item{second_diff.threshold}{Numeric value for ignoring small differences when using the "second_diff" threshold method.
May be useful for transformation that result in curves with small differences.
Generally, unlikely to be useful for most applications.
Default is 0.}

\item{chord.threshold}{Numeric value for ignoring small differences when using the "chord" threshold method.
May be useful for transformation that result in curves with small differences.
Generally, unlikely to be useful for most applications.
Default is 0.}
}
\value{
A \code{GRanges} object containing the classified regions and associated metadata of each,
including super enhancer status.
}
\description{
This function performs the ROSE for identifying super-enhancers by stitching
together peaks, calculating the signal in the regions, ranking the regions by signal,
and classifying them as super-enhancers.
}
\details{
This function allows for near identical functionality to the original ROSE software,
but also provides a number of additional options to minimize the impact of signal outliers
on the classification threshold.
In particular, the \code{transformation} argument allows for
the application of a function to the ranking signal before threshold determination.
This is highly recommended to prevent outliers from skewing the threshold.

The \code{thresh.method} argument allows for the selection of the method to determine the threshold.
\itemize{
\item The "ROSE" method is the default and uses a sliding diagonal line to determine the cutoff.
\item The "first" method uses a first derivative to determine the point where the slope is a given fraction of the maximum.
The \code{first.threshold} argument controls this fraction.
\item The "curvature" method finds the point at which the circle tanget to the curve has the smallest radius.
\item The "arbitrary" method allows for the user to specify a fixed threshold, useful for transformations
that result in a consistent curve shape with a known maximum, like cumulative proportion of signal.
}
}
\examples{
\dontrun{
sample.bam <- "path/to/sample.bam"
peaks <- "path/to/peaks.bed"
result <- run_rose(sample.bam, peaks)
}
}
\author{
Jared Andrews, Nicolas Peterson
}
